# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Attendance Intelligence API",
    "version": "0.1.0",
    "description": "Multi-tenant attendance intelligence: ingest attendance evidence, ask grounded questions with citations and confidence, export results, and improve answers through reviewer feedback. Every endpoint except health and the demo sign-in needs a bearer token (POST /api/v1/auth/dev-token in development). See README.md for examples."
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's tenant.",
              "title": "X-Tenant-Id"
            },
            "description": "Optional; must match the token's tenant."
          },
          {
            "name": "x-product-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's product.",
              "title": "X-Product-Id"
            },
            "description": "Optional; must match the token's product."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Feedback",
        "description": "Correct an answer. The example is validated and replayed; it is `active` (with its version and\nthe replayed answer as evidence) or `rejected` (with the reasons).",
        "operationId": "submit_feedback_api_v1_feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's tenant.",
              "title": "X-Tenant-Id"
            },
            "description": "Optional; must match the token's tenant."
          },
          {
            "name": "x-product-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's product.",
              "title": "X-Product-Id"
            },
            "description": "Optional; must match the token's product."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackExampleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List Feedback",
        "operationId": "list_feedback_api_v1_feedback_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "inactive",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's tenant.",
              "title": "X-Tenant-Id"
            },
            "description": "Optional; must match the token's tenant."
          },
          {
            "name": "x-product-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's product.",
              "title": "X-Product-Id"
            },
            "description": "Optional; must match the token's product."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback/{example_id}": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Get Feedback",
        "operationId": "get_feedback_api_v1_feedback__example_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "example_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Example Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's tenant.",
              "title": "X-Tenant-Id"
            },
            "description": "Optional; must match the token's tenant."
          },
          {
            "name": "x-product-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's product.",
              "title": "X-Product-Id"
            },
            "description": "Optional; must match the token's product."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackExampleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback/{example_id}/deactivate": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Deactivate Feedback",
        "description": "Roll an example back. It stops applying from the returned knowledge version, and cached answers\nmade with it are no longer served.",
        "operationId": "deactivate_feedback_api_v1_feedback__example_id__deactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "example_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Example Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's tenant.",
              "title": "X-Tenant-Id"
            },
            "description": "Optional; must match the token's tenant."
          },
          {
            "name": "x-product-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional; must match the token's product.",
              "title": "X-Product-Id"
            },
            "description": "Optional; must match the token's product."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackExampleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/dev-token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Token",
        "description": "Issue a token for a seeded demo user. Stands in for an identity provider in development.",
        "operationId": "dev_token_api_v1_auth_dev_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessContextResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "tenant_name": {
            "type": "string",
            "title": "Tenant Name"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "entity_scope": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entity Scope"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "clearance": {
            "type": "string",
            "title": "Clearance"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "departments": {
            "items": {
              "$ref": "#/components/schemas/DepartmentOut"
            },
            "type": "array",
            "title": "Departments"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "tenant_name",
          "product_id",
          "module",
          "user_id",
          "display_name",
          "role",
          "scope",
          "entity_scope",
          "employee_id",
          "clearance",
          "permissions",
          "departments"
        ],
        "title": "AccessContextResponse"
      },
...
      "ConfidenceOut": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "band": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Band"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "score",
          "band"
        ],
        "title": "ConfidenceOut"
      },
      "DepartmentOut": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "name"
        ],
        "title": "DepartmentOut"
      },
      "DevTokenRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._-]{2,63}$",
            "title": "User Id",
            "examples": [
              "acme.admin"
            ]
          },
          "product_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Product Id",
            "default": "hrms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "DevTokenRequest"
      },
...
      "FeedbackExampleOut": {
        "properties": {
          "example_id": {
            "type": "string",
            "title": "Example Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "rejected"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "The knowledge version at which the example was recorded or activated."
          },
          "scope": {
            "$ref": "#/components/schemas/FeedbackScopeOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "ideal_output": {
            "type": "string",
            "title": "Ideal Output"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "approval_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Note"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems",
            "description": "Why a rejected example was rejected.",
            "default": []
          },
          "replay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplayOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "deactivated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deactivated At"
          },
          "deactivated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deactivated By"
          },
          "knowledge_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Version",
            "description": "The scope's knowledge version after this change."
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "example_id",
          "status",
          "version",
          "scope",
          "request_id",
          "question",
          "feedback",
          "ideal_output",
          "reviewer_id",
          "created_at"
        ],
        "title": "FeedbackExampleOut"
      },
      "FeedbackListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedbackExampleOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FeedbackListOut"
      },
      "FeedbackRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{8,64}$",
            "title": "Request Id",
            "description": "The request_id of the answer being corrected."
          },
          "feedback": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Feedback",
            "description": "What was wrong, missing or undesirable."
          },
          "ideal_final_output": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Ideal Final Output",
            "description": "The answer the agent should give instead."
          },
          "approval_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Note",
            "description": "Why the reviewer approves it."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "feedback",
          "ideal_final_output"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackScopeOut": {
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "tenant",
              "department",
              "self"
            ],
            "title": "Scope"
          },
          "entity_scope": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entity Scope"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "clearance": {
            "type": "string",
            "title": "Clearance"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "entity_scope",
          "employee_id",
          "clearance"
        ],
        "title": "FeedbackScopeOut"
      },
      "FeedbackUsedOut": {
        "properties": {
          "example_id": {
            "type": "string",
            "title": "Example Id"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "The knowledge version at which the example was activated."
          },
          "similarity": {
            "type": "number",
            "title": "Similarity",
            "description": "How close this question is to the example's question (0-1)."
          }
        },
        "type": "object",
        "required": [
          "example_id",
          "version",
          "similarity"
        ],
        "title": "FeedbackUsedOut",
        "description": "An approved reviewer example that guided this answer."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplayOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity",
            "description": "Similarity of the replayed answer to the ideal output (0-1)."
          },
          "numbers_missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Numbers Missing",
            "default": []
          },
          "feedback_applied": {
            "items": {
              "$ref": "#/components/schemas/FeedbackUsedOut"
            },
            "type": "array",
            "title": "Feedback Applied",
            "default": []
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "outcome",
          "similarity"
        ],
        "title": "ReplayOut",
        "description": "The original question answered again with the example applied: the evidence for activation."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Seconds until the token expires."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}